+++ priv/repo/migrations/20250924215353_create_private_schema.exs
@@ -1,0 +1,13 @@
+defmodule TeslaMate.Repo.Migrations.CreatePrivateSchema do
+  use Ecto.Migration
+
+  def up do
+    execute("CREATE SCHEMA IF NOT EXISTS private;")
+    execute("ALTER TABLE public.tokens SET SCHEMA private;")
+  end
+
+  def down do
+    execute("ALTER TABLE private.tokens SET SCHEMA public;")
+    execute("DROP SCHEMA private;")
+  end
+end